+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/ExactTNCuda.cpp
@@ -12,13 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ObservablesTNCudaOperator.hpp"
 #include "ExactTNCuda.hpp"
-#include "MPSTNCuda.hpp"
 
-using namespace Pennylane::LightningTensor::TNCuda;
-
-template class Observables::ObservableTNCudaOperator<MPSTNCuda<float>>;
-template class Observables::ObservableTNCudaOperator<MPSTNCuda<double>>;
-template class Observables::ObservableTNCudaOperator<ExactTNCuda<float>>;
-template class Observables::ObservableTNCudaOperator<ExactTNCuda<double>>;
+// explicit instantiation
+template class Pennylane::LightningTensor::TNCuda::ExactTNCuda<float>;
+template class Pennylane::LightningTensor::TNCuda::ExactTNCuda<double>;
