+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/base/tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.20)
 
-project(tensornet_base_tests)
+project(tncuda_base_tests)
 
 # Default build type for test code is Debug
 if(NOT CMAKE_BUILD_TYPE)
@@ -14,25 +14,25 @@ FetchAndIncludeCatch()
 # Define library
 ################################################################################
 
-add_library(tensornet_base_tests INTERFACE)
-target_link_libraries(tensornet_base_tests INTERFACE  Catch2::Catch2)
+add_library(tncuda_base_tests INTERFACE)
+target_link_libraries(tncuda_base_tests INTERFACE  Catch2::Catch2)
 
-ProcessTestOptions(tensornet_base_tests)
+ProcessTestOptions(tncuda_base_tests)
 
 # Create dependency on the dynamically defined simulator/backend target.
-target_link_libraries(tensornet_base_tests INTERFACE ${PL_BACKEND} ${PL_BACKEND}_tncuda_utils)
+target_link_libraries(tncuda_base_tests INTERFACE ${PL_BACKEND} ${PL_BACKEND}_tncuda_utils)
 
-target_sources(tensornet_base_tests INTERFACE runner_lightning_tensor_tensornetBase.cpp)
+target_sources(tncuda_base_tests INTERFACE runner_lightning_tensor_TNCudaBase.cpp)
 
 ################################################################################
 # Define targets
 ################################################################################
-set(TEST_SOURCES    Tests_tensornetBase.cpp
+set(TEST_SOURCES    Tests_TNCudaBase.cpp
                     )
 
-add_executable(tensornet_base_test_runner ${TEST_SOURCES})
-target_link_libraries(tensornet_base_test_runner PRIVATE  tensornet_base_tests)
+add_executable(tncuda_base_test_runner ${TEST_SOURCES})
+target_link_libraries(tncuda_base_test_runner PRIVATE  tncuda_base_tests)
 
-catch_discover_tests(tensornet_base_test_runner)
+catch_discover_tests(tncuda_base_test_runner)
 
-install(TARGETS tensornet_base_test_runner DESTINATION bin)
+install(TARGETS tncuda_base_test_runner DESTINATION bin)
