+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/bindings/LTensorTNCudaBindings.hpp
@@ -138,8 +138,8 @@ void registerBackendClassSpecificBindings(PyClass &pyclass) {
                     py::buffer_info numpyArrayInfo = tensors[idx].request();
                     auto *data_ptr = static_cast<std::complex<PrecisionT> *>(
                         numpyArrayInfo.ptr);
-                    tensor_network.updateMPSSiteData(idx, data_ptr,
-                                                     tensors[idx].size());
+                    tensor_network.updateSiteData(idx, data_ptr,
+                                                  tensors[idx].size());
                 }
             },
             "Pass MPS site data to the C++ backend.")
