+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/gates/tests/CMakeLists.txt
@@ -15,10 +15,10 @@ FetchAndIncludeCatch()
 ################################################################################
 
 add_library(${PL_BACKEND}_gates_tests INTERFACE)
-target_link_libraries(${PL_BACKEND}_gates_tests INTERFACE     Catch2::Catch2
-                                                                ${PL_BACKEND}_gates
-                                                                ${PL_BACKEND}
-                                                                )
+target_link_libraries(${PL_BACKEND}_gates_tests INTERFACE Catch2::Catch2 
+                                                          ${PL_BACKEND}_gates
+                                                          ${PL_BACKEND}
+                                                          ${PL_BACKEND}_tncuda_utils)
 
 ProcessTestOptions(${PL_BACKEND}_gates_tests)
 
@@ -27,8 +27,9 @@ target_sources(${PL_BACKEND}_gates_tests INTERFACE runner_${PL_BACKEND}_gates.cp
 ################################################################################
 # Define targets
 ################################################################################
-set(TEST_SOURCES    Test_MPSTNCuda_NonParam.cpp
-                    Test_MPSTNCuda_Param.cpp)
+set(TEST_SOURCES Test_TNCuda_NonParam.cpp
+                 Test_TNCuda_Param.cpp
+                 Test_TNCuda_MPO.cpp)
 
 add_executable(${PL_BACKEND}_gates_test_runner ${TEST_SOURCES})
 target_link_libraries(${PL_BACKEND}_gates_test_runner PRIVATE ${PL_BACKEND}_gates_tests)
